requirements.txt
torchvision==0.8.1
torch==1.7.0
opencv_python==4.4.0.46
numpy==1.19.4
Pillow==8.1.0
